Move gridspec classes to gridspec.py
